refactor useWatcher functions
