+++ src/components/ConnectNode/index.tsx
@@ -35,7 +35,7 @@ export default function ConnectNode() {
   const handleLogout = () => {
     dispatch(authActions.resetState());
     dispatch(nodeActions.resetState());
-    dispatch(appActions.resetState());
+    dispatch(appActions.resetNodeState());
     dispatch(nodeActions.closeMessagesWebsocket());
     dispatch(nodeActions.closeLogsWebsocket());
     navigate('node/connect');
